System.OsString.Windows

Description

Deprecated: Use System.OsString.Posix from os-string >= 2.0.0 package instead. This module will be removed in filepath >= 1.5.

Types

data WindowsString

Commonly used windows string as wide character bytes.

data WindowsChar

String construction

encodeUtf :: MonadThrow m => String -> m WindowsString

Partial unicode friendly encoding.

This encodes as UTF16-LE (strictly), which is a pretty good guess.

Throws an EncodingException if encoding fails.

encodeWith :: TextEncoding -> String -> Either EncodingException WindowsString

Encode a String with the specified encoding.

encodeFS :: String -> IO WindowsString

This mimics the behavior of the base library when doing filesystem operations, which does permissive UTF-16 encoding, where coding errors generate Chars in the surrogate range.

The reason this is in IO is because it unifies with the Posix counterpart, which does require IO. This is safe to unsafePerformIO/unsafeDupablePerformIO.

fromBytes :: MonadThrow m => ByteString -> m WindowsString

Constructs a platform string from a ByteString.

This ensures valid UCS-2LE. Note that this doesn't expand Word8 to Word16 on windows, so you may get invalid UTF-16.

Throws EncodingException on invalid UCS-2LE (although unlikely).

pstr :: QuasiQuoter

QuasiQuote a WindowsString. This accepts Unicode characters and encodes as UTF-16LE on windows.

pack :: [WindowsChar] -> WindowsString

Pack a list of platform words to a platform string.

Note that using this in conjunction with unsafeFromChar to convert from [Char] to platform string is probably not what you want, because it will truncate unicode code points.

String deconstruction

decodeUtf :: MonadThrow m => WindowsString -> m String

Partial unicode friendly decoding.

This decodes as UTF16-LE (strictly), which is a pretty good.

Throws a EncodingException if decoding fails.

decodeWith :: TextEncoding -> WindowsString -> Either EncodingException String

Decode a WindowsString with the specified encoding.

The String is forced into memory to catch all exceptions.

decodeFS :: WindowsString -> IO String

Like decodeUtf, except this mimics the behavior of the base library when doing filesystem operations, which does permissive UTF-16 encoding, where coding errors generate Chars in the surrogate range.

The reason this is in IO is because it unifies with the Posix counterpart, which does require IO. unsafePerformIO/unsafeDupablePerformIO are safe, however.

unpack :: WindowsString -> [WindowsChar]

Unpack a platform string to a list of platform words.

Word construction

unsafeFromChar :: Char -> WindowsChar

Truncates to 2 octets.

Word deconstruction

toChar :: WindowsChar -> Char

Converts back to a unicode codepoint (total).