System.FilePath

Description

A library for FilePath manipulations, using Posix or Windows filepaths depending on the platform.

Both System.FilePath.Posix and System.FilePath.Windows provide the same interface. See either for examples and a list of the available functions.

Documentation

module System.FilePath.Posix