System.OsPath.Encoding

Types

data EncodingException

Constructors

EncodingError String (Maybe Word8)
Could not decode a byte sequence because it was invalid under the given encoding, or ran out of input in mid-decode.

showEncodingException :: EncodingException -> String

UCS-2

ucs2le :: TextEncoding

mkUcs2le :: CodingFailureMode -> TextEncoding

ucs2le_DF :: CodingFailureMode -> IO (TextDecoder ())

ucs2le_EF :: CodingFailureMode -> IO (TextEncoder ())

ucs2le_decode :: DecodeBuffer

ucs2le_encode :: EncodeBuffer

UTF-16LE_b

utf16le_b :: TextEncoding

Mimics the base encoding for filesystem operations. This should be total on all inputs (word16 byte arrays).

Note that this has a subtle difference to encodeWithBaseWindows/decodeWithBaseWindows: it doesn't care for the 0x0000 end marker and will as such produce different results. Use takeWhile (/= 'NUL') on the input to recover this behavior.

mkUTF16le_b :: CodingFailureMode -> TextEncoding

utf16le_b_DF :: CodingFailureMode -> IO (TextDecoder ())

utf16le_b_EF :: CodingFailureMode -> IO (TextEncoder ())

utf16le_b_decode :: DecodeBuffer

utf16le_b_encode :: EncodeBuffer

base encoding

encodeWithBasePosix :: String -> IO ShortByteString

This mimics the filepath dencoder base uses on unix, with the small distinction that we're not truncating at NUL bytes (because we're not at the outer FFI layer).

decodeWithBasePosix :: ShortByteString -> IO String

This mimics the filepath decoder base uses on unix, with the small distinction that we're not truncating at NUL bytes (because we're not at the outer FFI layer).

encodeWithBaseWindows :: String -> IO ShortByteString

This mimics the filepath dencoder base uses on windows, with the small distinction that we're not truncating at NUL bytes (because we're not at the outer FFI layer).

decodeWithBaseWindows :: ShortByteString -> IO String

This mimics the filepath decoder base uses on windows, with the small distinction that we're not truncating at NUL bytes (because we're not at the outer FFI layer).